+++ src/graphics/backend_wgpu/mod.rs
@@ -7,7 +7,7 @@ mod types;
 
 pub use font::Font;
 pub use quad::Quad;
-pub use surface::{winit, Surface};
+pub use surface::Surface;
 pub use texture::Texture;
 pub use triangle::Vertex;
 pub use types::TargetView;
@@ -19,6 +19,7 @@ use crate::{Error, Result};
 #[allow(missing_docs)]
 pub struct Gpu {
     device: wgpu::Device,
+    queue: wgpu::Queue,
     quad_pipeline: quad::Pipeline,
     triangle_pipeline: triangle::Pipeline,
     encoder: wgpu::CommandEncoder,
@@ -26,38 +27,50 @@ pub struct Gpu {
 
 impl Gpu {
     pub(super) fn for_window(
-        builder: winit::WindowBuilder,
-        events_loop: &winit::EventsLoop,
+        builder: winit::window::WindowBuilder,
+        event_loop: &winit::event_loop::EventLoop<()>,
     ) -> Result<(Gpu, Surface)> {
-        let instance = wgpu::Instance::new();
+        let window = builder
+            .build(event_loop)
+            .map_err(|error| Error::WindowCreation(error.to_string()))?;
 
-        let adapter = instance.get_adapter(&wgpu::AdapterDescriptor {
-            power_preference: wgpu::PowerPreference::HighPerformance,
+        let (mut device, queue) = futures::executor::block_on(async {
+            let adapter = wgpu::Adapter::request(
+                &wgpu::RequestAdapterOptions {
+                    power_preference: wgpu::PowerPreference::HighPerformance,
+                    compatible_surface: None,
+                },
+                wgpu::BackendBit::all(),
+            )
+            .await
+            .expect("Request adapter");
+
+            let (device, queue) = adapter
+                .request_device(&wgpu::DeviceDescriptor {
+                    extensions: wgpu::Extensions {
+                        anisotropic_filtering: false,
+                    },
+                    limits: wgpu::Limits::default(),
+                })
+                .await;
+
+            (device, queue)
         });
 
-        let mut device = adapter.request_device(&wgpu::DeviceDescriptor {
-            extensions: wgpu::Extensions {
-                anisotropic_filtering: false,
-            },
-            limits: wgpu::Limits::default(),
-        });
+        let surface = Surface::new(window, &device);
 
         let quad_pipeline = quad::Pipeline::new(&mut device);
         let triangle_pipeline = triangle::Pipeline::new(&mut device);
 
-        let window = builder
-            .build(events_loop)
-            .map_err(|error| Error::WindowCreation(error.to_string()))?;
-        let surface = Surface::new(window, &instance, &device);
-
         let encoder =
             device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                todo: 0,
+                label: Some("coffee::backend encoder"),
             });
 
         Ok((
             Gpu {
                 device,
+                queue,
                 quad_pipeline,
                 triangle_pipeline,
                 encoder,
@@ -67,7 +80,7 @@ impl Gpu {
     }
 
     pub(super) fn clear(&mut self, view: &TargetView, color: Color) {
-        let [r, g, b, a]: [f32; 4] = color.into_linear();
+        let [r, g, b, a] = color.into_linear();
 
         let _ = self.encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
             color_attachments: &[wgpu::RenderPassColorAttachmentDescriptor {
@@ -75,7 +88,12 @@ impl Gpu {
                 resolve_target: None,
                 load_op: wgpu::LoadOp::Clear,
                 store_op: wgpu::StoreOp::Store,
-                clear_color: wgpu::Color { r, g, b, a },
+                clear_color: wgpu::Color {
+                    r: r as f64,
+                    g: g as f64,
+                    b: b as f64,
+                    a: a as f64,
+                },
             }],
             depth_stencil_attachment: None,
         });
@@ -85,14 +103,19 @@ impl Gpu {
         &mut self,
         image: &image::DynamicImage,
     ) -> Texture {
-        Texture::new(&mut self.device, &self.quad_pipeline, image)
+        Texture::new(&mut self.device, &self.queue, &self.quad_pipeline, image)
     }
 
     pub(super) fn upload_texture_array(
         &mut self,
         layers: &[image::DynamicImage],
     ) -> Texture {
-        Texture::new_array(&mut self.device, &self.quad_pipeline, layers)
+        Texture::new_array(
+            &mut self.device,
+            &self.queue,
+            &self.quad_pipeline,
+            layers,
+        )
     }
 
     pub(super) fn create_drawable_texture(
@@ -102,6 +125,7 @@ impl Gpu {
     ) -> texture::Drawable {
         texture::Drawable::new(
             &mut self.device,
+            &self.queue,
             &self.quad_pipeline,
             width,
             height,
@@ -113,12 +137,14 @@ impl Gpu {
         drawable: &texture::Drawable,
     ) -> image::DynamicImage {
         let new_encoder = self.device.create_command_encoder(
-            &wgpu::CommandEncoderDescriptor { todo: 0 },
+            &wgpu::CommandEncoderDescriptor {
+                label: Some("coffee::backend encoder"),
+            },
         );
 
         let encoder = std::mem::replace(&mut self.encoder, new_encoder);
 
-        drawable.read_pixels(&mut self.device, encoder)
+        drawable.read_pixels(&mut self.device, &self.queue, encoder)
     }
 
     pub(super) fn upload_font(&mut self, bytes: &'static [u8]) -> Font {