+++ src/graphics/backend_wgpu/triangle.rs
@@ -1,6 +1,7 @@
 use std::mem;
 
 use crate::graphics::Transformation;
+use zerocopy::AsBytes;
 
 pub struct Pipeline {
     pipeline: wgpu::RenderPipeline,
@@ -17,24 +18,24 @@ impl Pipeline {
     pub fn new(device: &mut wgpu::Device) -> Pipeline {
         let transform_layout =
             device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-                bindings: &[wgpu::BindGroupLayoutBinding {
+                label: Some("coffee::backend::triangle transform"),
+                bindings: &[wgpu::BindGroupLayoutEntry {
                     binding: 0,
                     visibility: wgpu::ShaderStage::VERTEX,
-                    ty: wgpu::BindingType::UniformBuffer,
+                    ty: wgpu::BindingType::UniformBuffer { dynamic: false },
                 }],
             });
 
         let matrix: [f32; 16] = Transformation::identity().into();
 
-        let transform_buffer = device
-            .create_buffer_mapped(
-                16,
-                wgpu::BufferUsage::UNIFORM | wgpu::BufferUsage::TRANSFER_DST,
-            )
-            .fill_from_slice(&matrix[..]);
+        let transform_buffer = device.create_buffer_with_data(
+            matrix.as_bytes(),
+            wgpu::BufferUsage::UNIFORM | wgpu::BufferUsage::COPY_DST,
+        );
 
         let constant_bind_group =
             device.create_bind_group(&wgpu::BindGroupDescriptor {
+                label: Some("coffee::backend::triangle constants"),
                 layout: &transform_layout,
                 bindings: &[wgpu::Binding {
                     binding: 0,
@@ -50,29 +51,36 @@ impl Pipeline {
                 bind_group_layouts: &[&transform_layout],
             });
 
-        let vs_module = device
-            .create_shader_module(include_bytes!("shader/triangle.vert.spv"));
-        let fs_module = device
-            .create_shader_module(include_bytes!("shader/triangle.frag.spv"));
+        let vs = include_bytes!("shader/triangle.vert.spv");
+        let vs_module = device.create_shader_module(
+            &wgpu::read_spirv(std::io::Cursor::new(&vs[..]))
+                .expect("Read triangle vertex shader as SPIR-V"),
+        );
+
+        let fs = include_bytes!("shader/triangle.frag.spv");
+        let fs_module = device.create_shader_module(
+            &wgpu::read_spirv(std::io::Cursor::new(&fs[..]))
+                .expect("Read triangle fragment shader as SPIR-V"),
+        );
 
         let pipeline =
             device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
                 layout: &layout,
-                vertex_stage: wgpu::PipelineStageDescriptor {
+                vertex_stage: wgpu::ProgrammableStageDescriptor {
                     module: &vs_module,
                     entry_point: "main",
                 },
-                fragment_stage: Some(wgpu::PipelineStageDescriptor {
+                fragment_stage: Some(wgpu::ProgrammableStageDescriptor {
                     module: &fs_module,
                     entry_point: "main",
                 }),
-                rasterization_state: wgpu::RasterizationStateDescriptor {
+                rasterization_state: Some(wgpu::RasterizationStateDescriptor {
                     front_face: wgpu::FrontFace::Ccw,
                     cull_mode: wgpu::CullMode::None,
                     depth_bias: 0,
                     depth_bias_slope_scale: 0.0,
                     depth_bias_clamp: 0.0,
-                },
+                }),
                 primitive_topology: wgpu::PrimitiveTopology::TriangleList,
                 color_states: &[wgpu::ColorStateDescriptor {
                     format: wgpu::TextureFormat::Bgra8UnormSrgb,
@@ -89,36 +97,42 @@ impl Pipeline {
                     write_mask: wgpu::ColorWrite::ALL,
                 }],
                 depth_stencil_state: None,
-                index_format: wgpu::IndexFormat::Uint32,
-                vertex_buffers: &[wgpu::VertexBufferDescriptor {
-                    stride: mem::size_of::<Vertex>() as u64,
-                    step_mode: wgpu::InputStepMode::Vertex,
-                    attributes: &[
-                        wgpu::VertexAttributeDescriptor {
-                            shader_location: 0,
-                            format: wgpu::VertexFormat::Float2,
-                            offset: 0,
-                        },
-                        wgpu::VertexAttributeDescriptor {
-                            shader_location: 1,
-                            format: wgpu::VertexFormat::Float4,
-                            offset: 4 * 2,
-                        },
-                    ],
-                }],
+                vertex_state: wgpu::VertexStateDescriptor {
+                    index_format: wgpu::IndexFormat::Uint32,
+                    vertex_buffers: &[wgpu::VertexBufferDescriptor {
+                        stride: mem::size_of::<Vertex>() as u64,
+                        step_mode: wgpu::InputStepMode::Vertex,
+                        attributes: &[
+                            wgpu::VertexAttributeDescriptor {
+                                shader_location: 0,
+                                format: wgpu::VertexFormat::Float2,
+                                offset: 0,
+                            },
+                            wgpu::VertexAttributeDescriptor {
+                                shader_location: 1,
+                                format: wgpu::VertexFormat::Float4,
+                                offset: 4 * 2,
+                            },
+                        ],
+                    }],
+                },
                 sample_count: 1,
+                sample_mask: !0,
+                alpha_to_coverage_enabled: false,
             });
 
         let vertices = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("coffee::backend::triangle vertices"),
             size: mem::size_of::<Vertex>() as u64
                 * Self::INITIAL_BUFFER_SIZE as u64,
-            usage: wgpu::BufferUsage::VERTEX | wgpu::BufferUsage::TRANSFER_DST,
+            usage: wgpu::BufferUsage::VERTEX | wgpu::BufferUsage::COPY_DST,
         });
 
         let indices = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some("coffee::backend::triangle indices"),
             size: mem::size_of::<u32>() as u64
                 * Self::INITIAL_BUFFER_SIZE as u64,
-            usage: wgpu::BufferUsage::INDEX | wgpu::BufferUsage::TRANSFER_DST,
+            usage: wgpu::BufferUsage::INDEX | wgpu::BufferUsage::COPY_DST,
         });
 
         Pipeline {
@@ -146,9 +160,10 @@ impl Pipeline {
 
         let matrix: [f32; 16] = transformation.clone().into();
 
-        let transform_buffer = device
-            .create_buffer_mapped(16, wgpu::BufferUsage::TRANSFER_SRC)
-            .fill_from_slice(&matrix[..]);
+        let transform_buffer = device.create_buffer_with_data(
+            matrix.as_bytes(),
+            wgpu::BufferUsage::COPY_SRC,
+        );
 
         encoder.copy_buffer_to_buffer(
             &transform_buffer,
@@ -164,33 +179,29 @@ impl Pipeline {
             let new_size = vertices.len().max(indices.len()) as u32;
 
             self.vertices = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some("coffee::backend::triangle vertices"),
                 size: mem::size_of::<Vertex>() as u64 * new_size as u64,
-                usage: wgpu::BufferUsage::VERTEX
-                    | wgpu::BufferUsage::TRANSFER_DST,
+                usage: wgpu::BufferUsage::VERTEX | wgpu::BufferUsage::COPY_DST,
             });
 
             self.indices = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some("coffee::backend::triangle indices"),
                 size: mem::size_of::<u32>() as u64 * new_size as u64,
-                usage: wgpu::BufferUsage::INDEX
-                    | wgpu::BufferUsage::TRANSFER_DST,
+                usage: wgpu::BufferUsage::INDEX | wgpu::BufferUsage::COPY_DST,
             });
 
             self.buffer_size = new_size;
         }
 
-        let vertex_buffer = device
-            .create_buffer_mapped(
-                vertices.len(),
-                wgpu::BufferUsage::TRANSFER_SRC,
-            )
-            .fill_from_slice(vertices);
+        let vertex_buffer = device.create_buffer_with_data(
+            vertices.as_bytes(),
+            wgpu::BufferUsage::COPY_SRC,
+        );
 
-        let index_buffer = device
-            .create_buffer_mapped(
-                indices.len(),
-                wgpu::BufferUsage::TRANSFER_SRC,
-            )
-            .fill_from_slice(indices);
+        let index_buffer = device.create_buffer_with_data(
+            indices.as_bytes(),
+            wgpu::BufferUsage::COPY_SRC,
+        );
 
         encoder.copy_buffer_to_buffer(
             &vertex_buffer,
@@ -230,15 +241,16 @@ impl Pipeline {
 
             render_pass.set_pipeline(&self.pipeline);
             render_pass.set_bind_group(0, &self.constants, &[]);
-            render_pass.set_index_buffer(&self.indices, 0);
-            render_pass.set_vertex_buffers(&[(&self.vertices, 0)]);
+            render_pass.set_index_buffer(&self.indices, 0, 0);
+            render_pass.set_vertex_buffer(0, &self.vertices, 0, 0);
 
             render_pass.draw_indexed(0..indices.len() as u32, 0, 0..1);
         }
     }
 }
 
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, AsBytes)]
+#[repr(C)]
 pub struct Vertex {
     _position: [f32; 2],
     _color: [f32; 4],