+++ src/input/keyboard.rs
@@ -2,7 +2,7 @@
 
 mod event;
 
-pub use crate::graphics::window::winit::VirtualKeyCode as KeyCode;
+pub use crate::graphics::window::winit::event::VirtualKeyCode as KeyCode;
 pub use event::Event;
 
 use super::{ButtonState, Event as InputEvent, Input};