+++ src/input/mouse.rs
@@ -3,7 +3,7 @@
 mod event;
 mod wheel_movement;
 
-pub use crate::graphics::window::winit::MouseButton as Button;
+pub use crate::graphics::window::winit::event::MouseButton as Button;
 pub use event::Event;
 pub use wheel_movement::WheelMovement;