+++ src/ui/core/mouse_cursor.rs
@@ -23,15 +23,15 @@ pub enum MouseCursor {
 }
 
 #[doc(hidden)]
-impl From<MouseCursor> for winit::MouseCursor {
-    fn from(mouse_cursor: MouseCursor) -> winit::MouseCursor {
+impl From<MouseCursor> for winit::window::CursorIcon {
+    fn from(mouse_cursor: MouseCursor) -> winit::window::CursorIcon {
         match mouse_cursor {
-            MouseCursor::OutOfBounds => winit::MouseCursor::Default,
-            MouseCursor::Idle => winit::MouseCursor::Default,
-            MouseCursor::Pointer => winit::MouseCursor::Hand,
-            MouseCursor::Working => winit::MouseCursor::Progress,
-            MouseCursor::Grab => winit::MouseCursor::Grab,
-            MouseCursor::Grabbing => winit::MouseCursor::Grabbing,
+            MouseCursor::OutOfBounds => winit::window::CursorIcon::Default,
+            MouseCursor::Idle => winit::window::CursorIcon::Default,
+            MouseCursor::Pointer => winit::window::CursorIcon::Hand,
+            MouseCursor::Working => winit::window::CursorIcon::Progress,
+            MouseCursor::Grab => winit::window::CursorIcon::Grab,
+            MouseCursor::Grabbing => winit::window::CursorIcon::Grabbing,
         }
     }
 }